Remove incomprehensible Debian magic
I don't know how to fix this. It appears that debian/rules should
contain a override_dh_auto_install target, but I have not the slightest
clue what it should do or what it did with the old Makefile or how
to make it work with waf. Why the heck can't I just list the files
that should be installed?

Note: this is not a real removal. If someone wants to fix this, a
pull request reverting this commit + fixing the waf build would be
welcome. For now I'm removing this, because fixing it is probably
a huge time sink (just to learn Debian cryptomagic?), and I don't
want to confuse users with non-working instructions and packaging
files.