Fix bashism in debian/rules, handle errors better
POSIX shell syntax requires a ; before a closing ).
However, in this case the () wasn't even necessary since every
make command is issued in a shell of its own. Removed.
Replaced the ; after "cd mpv" with && so that errors are
detected properly.

While we're at it, make whitespace a bit more consistent throughout the file.