Add a simpler way to switch between release and master versions
This remembers the selection, and users are not tricked into
accidentally switching back to the release all the time.

Don't do this for libass, we always use master. (No reason to use a
usually buggy and outdated release.)

We don't do it for fribidi either, because fribidi is just in a
permanent state of bitrotting and brokenness.