Forcefully disable locale nonsense
Likes to break scripts. In this case, a russian user had a weird problem
with building mpv: the ffmpeg include files were not found. Maybe the
locale broke installing ffmpeg.

Adding this to every top-level script (annoyingly), but I won't add them
to the files in the scripts/ directory. The user normally doesn't need
to call them directly, so don't bother.