update script: use arguments normally instead of constructing commands
Previously $checkout_<proj> were constructed as commands which were
executed later. This is not very flexible and also hard to follow and
modify.

Change them to be used as normal arguments which contain the value
release/master, possibly with additional space-separated '-' to indicate
that the config file should be ignored (used when invoked with --master
or --release).