Remove Debian scripts
What’s so special about this distro that we not only have a repository
with scripts to compensate for the inability of its maintainers to
package and distribute software to our users, but also have to deal
with their incredibly bad tooling?

Also, the instructions don’t work and only waste users’ time. If these
“only exist for users who want to hurt themselves” then perhaps we
shouldn’t have them in the first place, even if “wants to hurt
themselves” practically describes the majority of Debian users.

Closes #106.