build options: allow CLI arguments with spaces
This affects CLI arguments to ./build, ./rebuild, as well as when
invoking any of scripts/*-build directly.

Previously, if such CLI argument contained a space, then the argument
was incorrectly split. Now spaces are taken correctly.

The issue was that the scripts used unquoted $@, which is identical to
unquoted $*, i.e. it's field-split on IFS.

Now we use quoted "$@", which preserves the arguments exactly.

Also, previously scripts/fribidi-build ignored the CLI arguments, and
now it doesn't (but we still override any -jN option with -j1).