configure/build: add meson support
This adds an optional environment variable, BUILDSYSTEM, that can be
passed to the build scripts to use mpv's meson build instead of waf,
the default. BUILDSYSTEM=meson needs to be set and any configure options
passed must use meson's syntax instead of waf. The catch with meson
support is that it requires the new prefer_static option for statically
linking libass and ffmpeg. This option is not currrently in a release
and will land in meson 0.63.