include libplacebo
Include libplacebo to add support `vo=gpu-next` in mpv,
as the OS-provided libplacebo may not be recent enough.

Explicitly link mpv and ffmpeg against stdc++ in case libplacebo was built with glslang,
which does not have pkg-config files.

Adjust the update script to also checkout git submodules as well,
which is needed for libplacebo.